% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cen.R
\name{cen}
\alias{cen}
\title{Centering matrix}
\usage{
cen(d)
}
\arguments{
\item{d}{integer giving the dimension of the matrix.}
}
\value{
A \code{numeric matrix} of dimension \code{d} by \code{d}.
}
\description{
Constructs a centering matrix of dimension \code{d} by \code{d}.
}
\details{
Using the centering matrix to center a matrix
is not computationally efficient.
}
\examples{
cen(1)
cen(2)
cen(3L)
cen(5)
cen(0) # Degenerate case works too
}
