% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cor}
\alias{cor}
\alias{cov}
\alias{pcor}
\alias{pcov}
\title{Compute the correlation matrix}
\usage{
cor(X)

cov(X, method = c("Unbiased", "ML"))
}
\arguments{
\item{X}{A numeric matrix with observations in rows and variables/features
in columns.}

\item{method}{A character of length 1. The unbiased estimate divided with
\code{n-1} and wheras ML uses \code{n}.}
}
\value{
\code{cor}: If \code{Y} is not supplied the a square symmetric correlation
  matrix of size \code{ncol(X)} is returned. If \code{Y} is given the
  cross-correlation matrix of size \code{ncol(X)} times \code{ncol(Y)} is
  returned.

\code{cov}: If \code{Y} is not supplied the a square symmetric covariance
  matrix of size \code{ncol(X)} is returned. If \code{Y} is given the
  cross-covariance matrix of size \code{ncol(X)} times \code{ncol(Y)} is
  returned.
}
\description{
This function offers the interface to compute the correlation
matrix between the columns of \code{X}.

This function offers the interface to compute the variance-covariance
matrix between the columns of \code{X}. If both \code{X} and \code{Y} is
supplied, the cross-covariances between the columns of the matrices are
computed.
}
\details{
Functions like \code{\link{cov}}. The \code{i}'th and \code{j}'th
  entry of the output matrix  is the correlation between \code{X[i, ]} and
  \code{Y[j, ]}. When \code{Y} is not supplied, then the output is equivalent
  to \code{Y} = \code{X}.
}
\examples{
X <- replicate(2, rnorm(10))
dimnames(X) <- list(paste0("obs", 1:nrow(X)), paste0("var", 1:ncol(X)))

cor(X)
cov(X, method = "ML")
cov(X, method = "Unbiased")
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link{corFamily}} (the workhorse functions)
}

