% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cor}
\alias{cor}
\alias{correlation}
\alias{cov}
\alias{covariance}
\alias{pcor}
\alias{pcov}
\alias{pxcor}
\alias{pxcov}
\alias{xcor}
\alias{xcov}
\title{Covariance and correlation}
\usage{
cor(X)

cov(X, method = c("Unbiased", "ML"))

pcor(X, z)

pcov(X, z)

pxcor(X, Y, Z)

pxcov(X, Y, Z)

xcor(X, Y)

xcov(X, Y, method = c("Unbiased", "ML"))
}
\arguments{
\item{X}{A numeric matrix with observations in rows and variables/features
in columns.}

\item{Y}{A numeric matrix with the same number of rows as \code{X}.}

\item{Z}{A numeric matrix with the same number of rows as \code{X}. This is
the}

\item{method}{A character of length 1. The unbiased estimate divided with
\code{n-1} and wheras ML uses \code{n}.}

\item{z}{A integer vector of indices to condition on.}
}
\value{
All functions return a matrix of correlations or covariances.
}
\description{
These functions offers a interface to compute arbitrary order partial
(or semi-partial) variance-covariance or correlation matrix as well as
partial cross variance-covariances or correlations.
}
\details{
Give some details!
}
\examples{
n <- 11
X <- createData(n, 4)
Y <- createData(n, 2)
Z <- createData(n, 9)

cor(X)
cov(X, method = "ML")
cov(X, method = "Unbiased")

xcov(X, Y)
xcor(X, Y)

pcov(X, z = numeric(0))  # == cov(X)
pcov(X, z = 1:3)
pcor(X, z = numeric(0))
pcor(X, z = 3)

pxcov(X, Y, Z)
pxcor(X, Y, Z)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link{corFamily}} (the workhorse functions)
}

