% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlateR-package.R
\docType{package}
\name{correlateR-package}
\alias{correlateR-package}
\alias{correlater}
\alias{correlateR}
\title{correlateR}
\description{
This R-package correlateR features fast, robust, and efficient marginal,
partial, semi-partial correlations and covariances of arbitrary conditional
order.
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
