% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/cov2cor.R
\name{cov2corArma}
\alias{cov2corArma}
\alias{cov2cor}
\title{Convert covariance matrix to correlation}
\usage{
cov2corArma(S)

cov2cor(S)
}
\arguments{
\item{S}{A square covariance matrix.}
}
\value{
A square correlation matrix.
}
\description{
This functions converts a covariance matrix \code{S} to a correlation matrix
fast and efficiently.
}
\examples{
X <- createData(n = 11, m = 4)
S <- cov(X)
stats::cov2cor(S)
cov2corArma(S)
if (require(microbenchmark)) {
  microbenchmark(A = cov2corArma(S),
                 B = stats::cov2cor(S),
                 C = cov2cor(S))
}
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
