% Generated by roxygen2 (4.0.1): do not edit by hand
\name{covariance}
\alias{covariance}
\title{Compute the covariance matrix}
\usage{
covariance(X, Y = NULL, method = c("Unbiased", "ML"))
}
\arguments{
\item{X}{A numeric matrix.}

\item{Y}{A numeric matrix of compatible dimension with the \code{X}, i.e.
\code{nrow(X)} equals \code{nrow(Y)}. The default value \code{NULL} is
equivalent to Y = X howver more efficient.}

\item{method}{A character of length 1. The unbiased estimate divided with
\code{n-1} and wheras ML uses \code{n}.}
}
\value{
If \code{Y} is not supplied the a square symmetric covariance matrix
  of size \code{ncol(X)} is returned. If \code{Y} is given the
  cross-covariance matrix of size \code{ncol(X)} times \code{ncol(Y)} is
  returned.
}
\description{
This function offers the interface to compute the variance-covariance
matrix between the columns of \code{X}. If both \code{X} and \code{Y} is
supplied, the cross-covariances between the columns of the matrices are
computed.
}
\details{
Functions like \link{\code{cov}}. The \code{i}'th and \code{j}'th
  entry of the output matrix  is the covariance between \code{X[i, ]} and
  \code{Y[j, ]}. When \code{Y} is not supplied, then the output is equivalent
  to \code{Y} = \code{X}.
}
\examples{
X <- replicate(2, rnorm(10))
dimnames(X) <- list(paste0("obs", 1:nrow(X)), paste0("var", 1:ncol(X)))
Y <- replicate(3, rnorm(10))
dimnames(Y) <- list(paste0("obs", 1:nrow(Y)), paste0("var", 1:ncol(Y)))

covariance(X)  # ==  covariance(X, X, "Unbiased")
covariance(X, method = "ML")
covariance(X, Y, "Unbiased")
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\link{\code{cov}}
}

