% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covRcpp}
\alias{covRcpp}
\alias{covFamily}
\alias{xcovRcpp}
\alias{covArma}
\alias{xcovArma}
\alias{covEigen}
\alias{xcovEigen}
\title{Marginal covariance matrix}
\usage{
covRcpp(X, norm_type = 0L)

xcovRcpp(X, Y, norm_type = 0L)

covArma(X, norm_type = 0L)

xcovArma(X, Y, norm_type = 0L)

covEigen(X, norm_type = 0L)

xcovEigen(X, Y, norm_type = 0L)
}
\arguments{
\item{X}{A numeric matrix.}

\item{norm_type}{an integer of length one giving the estimator. The 
default \code{0L} gives the unbiased estimate while \code{1L} gives the 
MLE.}

\item{Y}{A numeric matrix of compatible dimension with the \code{X}, i.e. 
\code{nrow(X)} equals \code{nrow(Y)}.}
}
\value{
The \code{cor}-family returns a numeric correlation matrix of size 
  \code{ncol(X)} times \code{ncol(X)}.
  
  The \code{xcor}-family returns a numeric cross-covariance matrix 
  of size \code{ncol(X)} times \code{ncol(Y)}.
}
\description{
Various workhorse functions to compute the marginal (or unconditional) 
covariance (and cross-covariance) estimates. The functions feature both the 
maximum likelihood and the biased corrected estimates. They are (almost) 
equivalent implementations of \code{\link[stats]{cov}} (\code{stats::cov}) 
in Rcpp, RcppArmadillo, and RcppEigen.
}
\details{
Functions almost like \code{\link{cor}}.
  For the \code{xcorXX} functions, the \code{i}'th and \code{j}'th 
  entry of the output matrix is the correlation between \code{X[i, ]} and 
  \code{X[j, ]}.
  Likewise, for the \code{xcorXX} functions, the \code{i}'th and
  \code{j}'th entry of the output is the correlation between \code{X[i, ]} 
  and \code{Y[j, ]}.
}
\note{
\code{NA}s in \code{X} or \code{Y} will yield NA's in the correlation matrix.
  This also includes the diagonal unlike the behavior of 
  \code{stats::cor(X)}.
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
