% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RCM.R
\name{createRCMData}
\alias{createRCMData}
\title{Simulate data from RCM}
\usage{
createRCMData(ns, psi, nu, sigma)
}
\arguments{
\item{ns}{A numeric vector giving the sample sizes in each study.}

\item{psi}{The underlying \eqn{Psi} parameter which is the mean
covariance matrix. If \code{nu} is \code{Inf}
this is used as \eqn{Sigma} parameter in all Wishart distributions.}

\item{nu}{A numeric of length one giving the underlying \eqn{nu} parameter.}

\item{sigma}{A numeric matrix giving the expected covariance matrix.
Can be supplied instead of \code{psi}.}
}
\value{
A \code{list} of matrices of the same size as \code{psi} giving
  observed scatter matrices from the RCM.\cr
  The realized covariance matrices are appended as an attribute.
}
\description{
Generate data from the hierarchical random covariance model (RCM).
}
\examples{
ns <- c(20, 10)
psi <- diag(3)
createRCMData(ns, psi, nu = 10)

createRCMData(ns, psi, nu = 1e20)
createRCMData(ns, psi, nu = Inf)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

