% Generated by roxygen2 (4.0.2): do not edit by hand
\name{difftest2.cor}
\alias{difftest2.cor}
\title{Test for difference in correlation}
\usage{
difftest2.cor(x1, y1, x2, y2, alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95)
}
\arguments{
\item{x1}{numeric vector, x-values for the first sample.}

\item{y1}{numeric vector, y-values for the first sample.}

\item{x2}{numeric vector, x-values for the second sample.}

\item{y2}{numeric vector, y-values for the second sample.}
}
\value{
A numeric vector giving correlation for each group,
  size-estimate and standard error, confidence intervals and p-values.
}
\description{
This functions uses Fisher's Z-transform (atanh) to test the null hypothesis
of no difference in correlations between x1 and y1 versus x2 and y2.
}
\details{
H0: cor(x1, y1) =  cor(x2, y2) \cr
two.sided => H1: cor(x1, y1) != cor(x2, y2) (i.e., z1 - z2 1= 0) \cr
greater   => H1: cor(x1, y1) >  cor(x2, y2) (i.e., z1 - z2 > 0) \cr
less      => H1: cor(x1, y1) <  cor(x2, y2) (i.e., z1 - z2 < 0)
}
\examples{
x1 <- rnorm(100)
y1 <- rnorm(100)
x2 <- rnorm(110)
y2 <- 4*x2 + 0.5*rnorm(110)

round(data.frame(
  two = correlateR:::difftest2.cor(x1, y1, x2, y2),
  les =  correlateR:::difftest2.cor(x1, y1, x2, y2, alternative = "less"),
  gre = correlateR:::difftest2.cor(x1, y1, x2, y2, alternative = "greater"),
  two2 = correlateR:::difftest2.cor(x2, y2, x1, y1)
  ),2)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link[stats]{cor.test}} in \code{stats}
}
\keyword{internal}

