% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RCM.R
\name{drcm}
\alias{drcm}
\title{Density of the RCM model}
\usage{
drcm(x, mu, Psi, nu, logarithm = FALSE)
}
\arguments{
\item{x}{A numeric matrix of size n by m with observations in the rows and
variables in the columns.}

\item{mu}{A numeric vector of length m giving the mean values of the RCM in
HDA.}

\item{Psi}{A numeric matrix giving the parameter of the RCM.}

\item{nu}{A numeric of length 1 giving the degrees of freedom.}

\item{logarithm}{A Boolean of length 1. If \code{TRUE}, the log density is
returned. Defaults to \code{FALSE}.}
}
\value{
Returns a vector of length n where the \eqn{i}th corresponds to the density
  evaluated in the \eqn{i}th row in \code{x}.
}
\description{
The density of the RCM model when the mean \code{mu} is non-zero. Used for
the HDA analysis.
}
\examples{
n <- 10
m <- 5
x <- matrix(rnorm(n*m), n, m)
mu <- rnorm(m)
Psi <- cov(matrix(rnorm(n*m), n, m))
nu <- 15
drcm(x, mu, Psi, nu)
drcm(x, mu, Psi, nu, logarithm = TRUE)
}

