% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fisher.test.cor}
\alias{fisher.test.cor}
\title{Fisher's correlation test}
\usage{
fisher.test.cor(estimate, mean, se, alternative, conf.level)
}
\arguments{
\item{estimate}{The fisher z-transformed correlation.}

\item{mean}{The fisher z-transformed null hypothess.}

\item{se}{The standard error of the estimate (usually \eqn{1/sqrt(n - 3)}
where n is the samplesize)}

\item{alternative}{specifies the alternative hypothesis.}

\item{conf.level}{confidence level for the returned confidence interval.}
}
\value{
A list of values resulting from the test.
}
\description{
Test hypothesis about the correlation using fisher's z-transform (atanh).
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\keyword{internal}

