% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fit.rcm}
\alias{fit.rcm}
\title{Fit using the EM algorithm}
\usage{
fit.rcm(S, ns, Psi.init = correlateR:::pool(S, ns), nu.init = sum(ns) + 1,
  max.ite = 1000, eps = 0.001, verbose = FALSE)
}
\arguments{
\item{S}{A \code{list} of square scatter matrices of the same size.}

\item{ns}{A vector of sample sizes corresponding to the scatter matrices in
\code{S}.}

\item{Psi.init}{A \code{matrix} giving the initial estimate of
\eqn{Psi}{Psi}.}

\item{nu.init}{A numeric of length one giving the inital estiamte of
\eqn{nu}{nu}.}

\item{max.ite}{A numeric of length one giving the maximum number of
iterations allowed.}

\item{eps}{The convergence criterion.}

\item{verbose}{If true, the differences in log-likelihood for each iteration
is printed out.}
}
\value{
A named list of length 3 with the elements:
  \item{Psi}{A matrix giving the estimate of \eqn{Psi}{Psi}.}
  \item{nu}{A number giving the estimate of \eqn{nu}{nu}.}
  \item{iterations}{A integer giving the number of iterations used.}
}
\description{
Fit the RCM using the modified EM algorithm.
}
\examples{
ns <- c(200, 150, 100)
Psi <- cor(createData(10, 5))
nu <- 30

S <- createRCMData(ns, Psi, nu)

print(res <- fit.rcm(S, ns, verbose = TRUE))
}
\seealso{
\code{\link{Psi2Sigma}}
}

