% Generated by roxygen2 (4.0.2): do not edit by hand
\name{grem_get_nu}
\alias{grem_get_nu}
\title{Estimate degrees of freedom}
\usage{
grem_get_nu(Psi, nu, S, ns, interval)
}
\arguments{
\item{Psi}{A numeric matrix of size \eqn{p} times \eqn{p} giving the initial
estimate of \eqn{Psi}{\Psi}.}

\item{nu}{A single numeric number giving the degrees of freedom
\eqn{nu}{\nu}.}

\item{S}{A \code{list} of scatter matrices.}

\item{ns}{Vector of group sizes.}
}
\value{
A single number giving the \eqn{nu}{\nu} optimizing the GREM
  likelihood with fixed \eqn{Psi}{\Psi}.
}
\description{
Function for estimating the degrees of freedom \eqn{nu}{\nu} in the
Graphical Random Effects Model (GREM).
}
\author{
Anders Ellern Bilgrau
}
\keyword{internal}

