% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ICC}
\alias{ICC}
\title{Compute the ICC in the RCM}
\usage{
ICC(nu, p)
}
\arguments{
\item{nu}{A numeric giving the estimated degrees of freedom.}

\item{p}{A numeric giving the dimension of the space.}
}
\value{
A numeric giving the ICC.
}
\description{
A simple function for computing the intra-class correlation coefficient
(ICC). This function simply computes \code{1/(nu - p)}.
}

