% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Psi2Sigma}
\alias{Psi2Sigma}
\title{Conversion from Psi and nu to Sigma}
\usage{
Psi2Sigma(Psi, nu)
}
\arguments{
\item{Psi}{A numeric square positive semi-definite matrix.}

\item{nu}{A numeric of length one giving the degrees of freedom in the RCM.}
}
\value{
A covariance matrix the same size as \code{Psi}.
}
\description{
Computes the expected covariance matrix from Psi and nu in the random
covariance model (RCM).
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

