% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pxcor}
\alias{pxcor}
\alias{pxcov}
\alias{xcor}
\alias{xcov}
\title{Partial cross covariance and correlation}
\usage{
pxcor(X, Y, Z)

pxcov(X, Y, Z)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{y}{A numeric vector of observations.}

\item{Z}{A numeric matrix of observations to condition on.}
}
\value{
\code{pxcor} retuns a numeric  matrix of partial cross correlations
  between \code{X} and \code{Y} given \code{Z}.
library(RcppEigen)
n <- 40
X <- replicate(4, rnorm(n))
Y <- replicate(2, rnorm(n))
Z <- replicate(11, rnorm(n))
xcor(X, Y, Z)

\code{pxcov} retuns a numeric  matrix of partial cross covariances
  between \code{X} and \code{Y} given \code{Z}.
}
\description{
Compute the partial cross (x) correlation or covariance using linear
regression
}

