% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcm_em_step_arma}
\alias{rcm_em_step_arma}
\title{The RCM EM-step}
\usage{
rcm_em_step_arma(Psi, nu, S_list, ns)
}
\arguments{
\item{Psi}{A numeric matrix.}

\item{nu}{A numeric of length 1 giving the degrees of freedom in the RCM.}

\item{S_list}{A \code{list} of scatter matrices for each dataset/group
of the same size a \code{Psi}.}

\item{ns}{A numeric vector the same lengths as \code{S_list} giving the
number of samples for each dataset.}
}
\value{
A numeric matrix the same size as \code{Psi} giving the updated
  \code{Psi}.
}
\description{
A armadillo-based function to perform the E and M step in the
EM algorithm of the RCM. This functions assumes \code{nu} to be fixed.
}
\examples{
ns <-  c(5, 5, 5)
S <- createRCMData(ns = ns, psi = diag(4), nu = 30)
correlateR:::rcm_em_step_arma(Psi = diag(4), nu = 15, S_list = S, ns = ns)
}
\keyword{internal}

