% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rcm_get_nu}
\alias{rcm_get_nu}
\alias{rcm_get_nu2}
\title{Estimate degrees of freedom}
\usage{
rcm_get_nu(Psi, S_list, ns)

rcm_get_nu2(Psi, S_list, ns)
}
\arguments{
\item{Psi}{A numeric matrix of size \eqn{p} times \eqn{p} giving the initial
estimate of \eqn{Psi}{\Psi}.}

\item{S_list}{A \code{list} of scatter matrices.}

\item{ns}{Vector of group sizes.}
}
\value{
A list giving the \eqn{nu}{\nu} optimizing the RCM
  likelihood with fixed \eqn{Psi}{\Psi} and other stuff.
}
\description{
Function for estimating the degrees of freedom \eqn{nu}{\nu} in the
random covariance model (RCM).
}
\note{
\code{rcm_get_nu} optimizes via \code{\link{optimize}}.

\code{rcm_get_nu2} optimizes via \code{\link{nlm}}.
}
\examples{
p <- 3
Psi <- diag(p)
ns <- c(5, 5, 5)
true.nu <- 7
nus <- seq(p+1+1e-5, 10, by = 0.01)
S_list <- createRCMData(ns = ns, psi = Psi, nu = true.nu)
eval.ll <- c()
for (i in seq_along(nus)) {
  eval.ll[i] <- correlateR:::rcm_loglik_nu_arma(Psi, nus[i], S_list, ns)
}
plot(nus, eval.ll, type = "l", ylim = c(-30, 0))
abline(v = true.nu, col = "red", lwd = 2)

# Get nu
print(ans <- correlateR:::rcm_get_nu(Psi, S_list, ns))
print(ans2 <- correlateR:::rcm_get_nu2(Psi, S_list, ns))

abline(v = ans$maximum, col = "orange", lwd = 2, lty = 2)
abline(v = ans2$estimate, col = "blue", lwd = 2. lty = 3)

\dontrun{
library("microbenchmark")
microbenchmark(correlateR:::rcm_get_nu(Psi, S_list, ns),
               correlateR:::rcm_get_nu2(Psi, S_list, ns))
}
}
\author{
Anders Ellern Bilgrau
}
\keyword{internal}

