% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rcm_loglik_arma}
\alias{rcm_loglik_arma}
\title{The RCM log-likelihood function}
\usage{
rcm_loglik_arma(Psi, nu, S_list, ns)
}
\arguments{
\item{Psi}{A numeric matrix.}

\item{nu}{A numeric of length 1 giving the degrees of freedom.}

\item{S_list}{A \code{list} of scatter matrices of the same size as
\code{Psi} for each group.}

\item{ns}{A numeric of the same length as \code{S_list} giving the
number of samples in each group.}
}
\value{
The value of the log-likelihood.
}
\description{
Fast evaluation of the RCM log-likelihood function.
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\keyword{internal}

