% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Scov}
\alias{Scov}
\title{Shrinkage covariance estimation}
\usage{
Scov(X, method = c("RBLW", "LW", "OAS"))
}
\arguments{
\item{X}{The data matrix of size \code{n} by \code{p}.}

\item{method}{The method of estimating the optimal interpolating parameter.}
}
\value{
A \code{p} by \code{p} numeric matrix with two extra attributes giving
  the used mixture (\eqn{\rho}{rho}) and the method.
}
\description{
This algorithm seeks to find a covariance estimate that (asymptotically)
minimizes the mean-squared error (MSE) obtained by shrinkage as proposed by
Leodit-Wolf (LW). It is effectively a interpolation/mix of the sample ML
estimate of the covariance matrix, \eqn{S}, and the most well-conditioned
(and naive) estimate \eqn{F = 1/p tr(S) I}.
The algorithm seeks a solution to the problem:
  \deqn{\min_{\rho}E[ || \Sigma_O - \Sigma ||^2 ]}{
        mininimize E[ || Sigma_O - Sigma ||^2 ] w.r.t. rho}
  \deqn{\text{s.t.} \Sigma_O = (1 - \rho)S + \rho F}{
        s.t. Sigma_O = (1-rho)*S + rho*F}
}
\details{
The improved estimate using Rao-Blackwell theorem, abbreviated RBLW, is
  also implemented.
  The intepolated \eqn{\rho}{rho} value used is always
  \eqn{\min(\rho, 1)}{min(rho,1)}.
}
\examples{
n <- 3
X <- createData(n, 5)
cov(X)
Scov(X, method = "RBLW")
Scov(X, method = "LW")
}
\references{
Chen, Y., Member, S. S., Wiesel, A., Eldar, Y. C., Member, S. S., &
  Hero, A. O. (2009). Shrinkage Algorithms for MMSE Covariance Estimation,
  58(734), 1–28. Methodology; Computation.
  Retrieved from http://arxiv.org/abs/0907.4698
}

