% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SimulateRCMData}
\alias{SimulateRCMData}
\title{Simulate data from RCM}
\usage{
SimulateRCMData(ns, psi, nu)
}
\arguments{
\item{ns}{A numeric vector giving the sample sizes in each study.}

\item{psi}{The underlying \eqn{Psi} parameter. If \code{nu} is \code{Inf}
this is used as \eqn{Sigma} parameter in all Wishart distribution.}

\item{nu}{A numeric of length one giving the underlying \eqn{nu} parameter.}
}
\value{
A \code{list} of matrices of the same size as \code{psi} giving
  observed scatter matrices from the RCM.
}
\description{
Generate data from the hierarchical random covariance model (RCM).
}
\examples{
ns <- c(20, 14, 10)
psi <- diag(3)
SimulateRCMData(ns, psi, nu = 10)

SimulateRCMData(ns, psi, nu = 1e20)
# is NOT the same as
SimulateRCMData(ns, psi, nu = Inf)
# which is almost the same as
SimulateRCMData(ns, psi*(1e20 - 3 - 1), nu = 1e20)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

