% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{The soft thresholding function}
\usage{
soft(a, c)
}
\arguments{
\item{a}{A \code{numeric} vector or \code{matrix} to apply the soft 
thresholding function on.}

\item{c}{A \code{numeric} giving the soft threshold.}
}
\value{
A \code{numeric} vector or \code{matrix} the same size as \code{a}
   with thresholded values.
}
\description{
Element-wise application of the soft thresholding function given by
\deqn{soft(a, c) = sgn(a)\cdot(|a| - c)_+}{
      soft(a, c) = sgn(a)*max(|a| - c, 0)}
}
\examples{
a <- rnorm(100)
soft(a, c = 0.2)
  
x <- createData(n = 10, m = 4)
soft(x, c = 0.7)
}
