% Generated by roxygen2 (4.0.2): do not edit by hand
\name{test.xcor}
\alias{test.xcor}
\title{Test for correlation between multiple variables samples}
\usage{
\method{test}{xcor}(X, Y, alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95, null = 0)
}
\arguments{
\item{X}{numeric matrix or vector of data values.}

\item{Y}{numeric matrix or vector of data values.
If matrices, \code{X} and \code{Y} must have the same number of rows.
If vectors, \code{X} and \code{Y} mush have the same length.}

\item{alternative}{specifies the alternative hypothesis.}

\item{conf.level}{confidence level for the returned confidence interval.}

\item{null}{The null to test against. Either a single number of or a matrix
of size \code{ncol(X)} times \code{ncol(Y)}.}
}
\value{
If matrices are given a list of matrices are returned where the \eqn{ij}th
  entry corresponds to the results of \code{stats::cor.test(X[,i], Y[,j])}.\cr
  If vectors are given a vector of the results is given.
}
\description{
Test for Pearson correlation between paired samples across many variables.
}
\details{
The test and confidence interval is based Fisher's Z transform.
}
\examples{
X <- createData(10, 2)
Y <- createData(10, 4)
test.xcor(X, Y)
test.xcor(X[,1], Y[,1])
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link[stats]{cor.test}}
}

