% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/arbitraryOrderCorrelation.R
\name{arbitraryOrderCorrelation}
\alias{arbitraryOrderCorrelation}
\title{Partial correlations}
\usage{
arbitraryOrderCorrelation(S, z)
}
\arguments{
\item{S}{A numeric correlation matrix}

\item{z}{A integer vector of indices of the variables to condition on. I.e.
length(z) is the order of the partial correlations. If \code{z} has
length 0 the marginal correlations are returned.}
}
\value{
A numeric matrix of the same size as the correlation matrix
  with the partial correlations given the variables indexed by \code{z}.
  \code{NaN} are returned in the rows and columns of the conditioned
  variables.
}
\description{
Compute arbitrary \code{k}th order partial correlations using recursion.
This implementation is very inefficient.
}
\examples{
X <- createData(6, 20)
Y <- createData(6, 20)
Z <- createData(6, 20)
S <- cor(X)
arbitraryOrderCorrelation(S, z = c(2, 3, 5))
}
\seealso{
\code{\link{correlation}} \code{\link{correlation}}
}

