% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R, R/cov.R, R/pcor.R, R/pcov.R, R/pxcor.R,
%   R/pxcov.R, R/xcor.R, R/xcov.R
\name{cor}
\alias{cor}
\alias{cov}
\alias{pcor}
\alias{pcov}
\alias{xcor}
\alias{xcov}
\alias{pxcor}
\alias{pxcov}
\alias{correlation}
\alias{covariance}
\title{Covariance and correlation}
\usage{
cor(X)

cov(X, method = c("Unbiased", "ML"))

pcor(X, z)

pcov(X, z)

pxcor(X, Y, Z)

pxcov(X, Y, Z, method = c("Unbiased", "ML"))

xcor(X, Y)

xcov(X, Y, method = c("Unbiased", "ML"))
}
\arguments{
\item{X}{A numeric matrix with observations in rows and variables
in columns.}

\item{method}{A character of length 1. The unbiased estimate divided with
\code{n-1} and whereas ML uses \code{n}.}

\item{z}{A integer vector of indices to condition on.}

\item{Y}{A numeric matrix with the same number of rows as \code{X}.}

\item{Z}{A numeric matrix with the same number of rows as \code{X}. This is
the matrix of variables to condition on.}
}
\value{
All functions return a matrix of correlations or covariances.
}
\description{
These functions offer an interface to compute arbitrary order partial
(or semi-partial) variance-covariance or correlation matrices as well as
partial cross variance-covariances or correlations.
}
\examples{
n <- 11
X <- createData(n, 4)
Y <- createData(n, 2)
Z <- createData(n, 9)

cor(X)
cov(X, method = "ML")
cov(X, method = "Unbiased")

xcov(X, Y)
xcor(X, Y)

pcov(X, z = numeric(0)) 
pcov(X, z = 1:2)
pcor(X, z = numeric(0))
pcor(X, z = 3)

pxcov(X, Y, Z)
pxcor(X, Y, Z)
}
\seealso{
\code{\link{corFamily}} (the internal workhorse functions)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
