% Generated by roxygen2 (4.0.2): do not edit by hand
\name{corRcpp}
\alias{corArma}
\alias{corEigen}
\alias{corFamily}
\alias{corRcpp}
\alias{xcorArma}
\alias{xcorEigen}
\alias{xcorRcpp}
\title{Marginal correlation matrix}
\usage{
corRcpp(X)

xcorRcpp(X, Y)

corArma(X)

xcorArma(X, Y)

corEigen(X)

xcorEigen(X, Y)
}
\arguments{
\item{X}{A numeric matrix.}

\item{Y}{A numeric matrix of compatible dimension with the \code{X}, i.e.
\code{nrow(X)} equals \code{nrow(Y)}.}
}
\value{
The \code{corXX} familiy returns a numeric correlation matrix of size
  \code{ncol(X)} times \code{ncol(X)}.

  The \code{xcorXX} family returns a numeric cross-correlation matrix
  of size \code{ncol(X)} times \code{ncol(Y)}.
}
\description{
Various workhorse functions to compute the marginal (or unconditional)
correlations (and cross-correlation) estimates efficiently.
They are (almost)
equivalent implementations of \code{\link[stats]{cor}} in Rcpp,
RcppArmadillo, and RcppEigen.
}
\details{
Functions almost like \code{\link{cor}}.
  For the \code{xcorXX} functions, the \code{i}'th and \code{j}'th
  entry of the output matrix is the correlation between \code{X[i, ]} and
  \code{X[j, ]}.
  Likewise, for the \code{xcorXX} functions, the \code{i}'th and
  \code{j}'th entry of the output is the correlation between \code{X[i, ]}
  and \code{Y[j, ]}.
}
\note{
NA's in \code{X} or \code{Y} will yield NA's in the correlation matrix.
  This also includes the diagonal unlike the behaviour of
  \code{\link[stats]{cor}}.
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

