% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cov2cor}
\alias{cov2cor}
\alias{cov2corArma}
\title{Convert covariance matrix to correlation}
\usage{
cov2cor(S)

cov2corArma(S)
}
\arguments{
\item{S}{A square covariance matrix.}
}
\value{
A square correlation matrix.
}
\description{
This functions converts a covariance matrix \code{S} to a correlation matrix
fast and efficiently.
}
\examples{
X <- createData(n = 11, m = 4)
S <- cov(X)
stats::cov2cor(S)
cov2corArma(S)
if (require(microbenchmark)) {
  microbenchmark(A = cov2corArma(S),
                 B = stats::cov2cor(S),
                 C = cov2cor(S))
}
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

