% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createData}
\alias{createData}
\title{Create toy data}
\usage{
createData(n, m, n.na = 0)
}
\arguments{
\item{n}{an integer giving the number of observations/samples.}

\item{m}{an integer giving the number of variables/features.}

\item{n.na}{an integer giving the number of \code{NA}s in the output.
Default is \code{0}.}
}
\value{
A \code{n} times \code{m} numeric matrix of observations.
}
\description{
Creates a named matrix of toy data from an \code{m}-dimensional
zero-mean normal distribution with identity covariance matrix.
}
\note{
The \code{NA} are randomly inserted in the output.
}
\examples{
createData(10, 4)
}

