% Generated by roxygen2 (4.0.2): do not edit by hand
\name{firstOrderPartialCorrelation}
\alias{firstOrderPartialCorrelation}
\title{Compute first order partial correlations}
\usage{
firstOrderPartialCorrelation(S, k)
}
\arguments{
\item{S}{A numeric correlation matrix}

\item{k}{An integer. The index of the variable to condition on.}
}
\value{
A matrix of the same size as \code{X} with \code{NaN}s in the
  \code{k} row and column.
}
\description{
This function computes the first order partial correlation given the
\code{k}'th variable.
}
\examples{
X <- createData(4, 10)
S <- cor(X)
firstOrderPartialCorrelation(S, 2)
}
\author{
Anders Ellernsta Bilgrau <abilgrau (at) math.aau.dk>
}
\seealso{
\code{\link{cor}}
}

