% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RCM.R
\name{fit.rcm}
\alias{fit.rcm}
\title{Fit using the EM algorithm}
\usage{
fit.rcm(S, ns, Psi.init, nu.init, method = c("EM", "pooled", "approxMLE"),
  conf.lvl = 0.95, max.ite = 1000, eps = 0.001, verbose = FALSE)
}
\arguments{
\item{S}{A \code{list} of square scatter matrices of the same size.}

\item{ns}{A vector of group or sample sizes corresponding to the scatter
matrices in \code{S}.}

\item{Psi.init}{A \code{matrix} giving the initial estimate of
\eqn{Psi}{Psi}. Default starting value is the scaled pooled sample
covariance matrix.}

\item{nu.init}{A numeric of length one giving the initial estimate of
\eqn{nu}{nu}. Default is \code{sum(ns)}.}

\item{method}{A character giving the method to be used or abbreviation
hereof.}

\item{conf.lvl}{The confidence level. Default is 0.95.}

\item{max.ite}{A numeric of length one giving the maximum number of
iterations allowed. Default is 1000.}

\item{eps}{The convergence criterion.}

\item{verbose}{If true, the differences in log-likelihood for each iteration
is printed out.}
}
\value{
A named list of length 3 with the elements:
  \item{Psi}{A matrix giving the estimate of \eqn{Psi}{Psi}.}
  \item{nu}{A number giving the estimate of \eqn{nu}{nu}.}
  \item{iterations}{A integer giving the number of iterations used.}
  \item{loglik}{A numeric giving the value of the log-likelihood in the last
     iteration.}
}
\description{
Fit the RCM using the modified EM algorithm.
}
\examples{
nss <- c(40, 20, 30, 10)
print(Psii <- drop(rwishart(1)))  # Expected covariance
nuu <- 7
SS <- createRCMData(ns = nss, psi = Psii, nu = nuu)

fit.rcm(SS, nss, method = "EM", verbose = TRUE)
fit.rcm(SS, nss, method = "pool", verbose = TRUE)
fit.rcm(SS, nss, method = "approxMLE", verbose = FALSE)
}

