% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fit.rcm}
\alias{fit.rcm}
\title{Fit using the EM algorithm}
\usage{
fit.rcm(S, ns, Psi.init, nu.init, method = c("EM", "pool", "approxMLE"),
  max.ite = 1000, eps = 0.001, verbose = FALSE)
}
\arguments{
\item{S}{A \code{list} of square scatter matrices of the same size.}

\item{ns}{A vector of group or sample sizes corresponding to the scatter
matrices in \code{S}.}

\item{Psi.init}{A \code{matrix} giving the initial estimate of
\eqn{Psi}{Psi}. Default starting value is the scaled pooled sample
covariance matrix.}

\item{nu.init}{A numeric of length one giving the inital estimate of
\eqn{nu}{nu}. Default is \code{sum(ns)}.}

\item{method}{The method to be used.}

\item{max.ite}{A numeric of length one giving the maximum number of
iterations allowed. Default is 1000.}

\item{eps}{The convergence criterion.}

\item{verbose}{If true, the differences in log-likelihood for each iteration
is printed out.}
}
\value{
A named list of length 3 with the elements:
  \item{Psi}{A matrix giving the estimate of \eqn{Psi}{Psi}.}
  \item{nu}{A number giving the estimate of \eqn{nu}{nu}.}
  \item{iterations}{A integer giving the number of iterations used.}
}
\description{
Fit the RCM using the modified EM algorithm.
}
\examples{
ns <- rep(11, 3)
print(Psi <- drop(rwishart(1)))
nu <- 30
S <- createRCMData(ns, Psi, nu)

with(fit.rcm(S, ns, method = "EM"),        Psi2Sigma(Psi, nu))
with(fit.rcm(S, ns, method = "pool"),      Psi2Sigma(Psi, nu))
with(fit.rcm(S, ns, method = "approxMLE"), Psi2Sigma(Psi, nu))
Psi2Sigma(Psi, nu)
}
\seealso{
\code{\link{Psi2Sigma}}
}

