% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pool}
\alias{pool}
\title{Pooled covariance from list of scatter matrices}
\usage{
pool(S_list, ns, norm_type = 0L)
}
\arguments{
\item{S_list}{A list of scatter matrices.}

\item{norm_type}{A integer of length one equaling \code{0} or \code{1}. 
If \code{0} the bias correction is used. If \code{1} the ML estimate is 
used.}

\item{nu}{A numeric vector giving the number of samples corresponding
to each scatter matrix.}
}
\value{
A numeric matrix giving the pooled variance.
}
\description{
Pooled covariance from list of scatter matrices
}
\author{
Anders Ellern Bilgrau
}
\keyword{internal}
