% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Psi2Sigma}
\alias{Psi2Sigma}
\alias{Sigma2Psi}
\title{Conversion from Psi between and Sigma}
\usage{
Psi2Sigma(Psi, nu)

Sigma2Psi(Sigma, nu)
}
\arguments{
\item{Psi}{A numeric square positive semi-definite matrix. The underlying
parameter in the RCM.}

\item{nu}{A numeric of length one giving the degrees of freedom in the RCM.}

\item{Sigma}{A numeric square positive semi-definite matrix. The expected
covariance matrix in the RCM.}
}
\value{
The converted matrix the same size as \code{Psi} or \code{Sigma}.
}
\description{
Computes the expected covariance matrix from Psi and nu in the random
covariance model (RCM) and vice versa.
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}

