% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RCM.R
\name{rcm_get_nu}
\alias{rcm_get_nu}
\alias{rcm_get_nu_nlm}
\alias{rcm_get_nu_optim}
\alias{rcm_get_nu_optimize}
\alias{rcm_get_nu_optimize2}
\title{Estimate degrees of freedom}
\usage{
rcm_get_nu(Psi, S_list, ns)

rcm_get_nu_optimize(Psi, S_list, ns)

rcm_get_nu_optim(Psi, S_list, ns, ...)

rcm_get_nu_nlm(Psi, S_list, ns, ...)

rcm_get_nu_optimize2(Psi, S_list, ns)
}
\arguments{
\item{Psi}{A numeric matrix of size \eqn{p} times \eqn{p} giving the initial
estimate of \eqn{Psi}{\Psi}.}

\item{S_list}{A \code{list} of scatter matrices.}

\item{ns}{Vector of group sizes.}

\item{\dots}{arguments passed to the optimizer.}
}
\value{
A list giving the \eqn{nu}{\nu} optimizing the RCM
  likelihood with fixed \eqn{Psi}{\Psi} and other stuff.
}
\description{
Functions for estimating the degrees of freedom \eqn{nu}{\nu} in the
random covariance model (RCM).
}
\note{
\code{rcm_get_nu} is a wrapper for \code{rcm_get_nu_optimize}.

\code{rcm_get_nu_optimize} optimizes via \code{\link{optimize}} and
  the Brent-type optimization.

\code{rcm_get_nu_optim} optimizes via \code{\link{optim}} and the
  L-BFGS-B method.

\code{rcm_get_nu_nlm} optimizes via \code{\link{nlm}}.

\code{rcm_get_nu_optim2} optimizes via \code{\link{optimize}} and the
  Brent-type method. A faster implementation. Avoids many repeated
  evaluations.
}
\examples{
p <- 3
Psi <- diag(p)
ns <- c(5, 5, 5)
true.nu <- 7
nus <- seq(p - 1 + sqrt(.Machine$double.eps), 40, l = 1000)
S_list <- createRCMData(ns = ns, psi = Psi, nu = true.nu)
eval.ll <- c()
for (i in seq_along(nus)) {
  eval.ll[i] <- correlateR:::rcm_loglik_arma(Psi, nus[i], S_list, ns)
}
plot(nus, eval.ll, type = "l")
abline(v = true.nu, col = "red", lwd = 2)

# Get nu
print(ans <- correlateR:::rcm_get_nu_optimize(Psi, S_list, ns))
print(ans2 <- correlateR:::rcm_get_nu_optim(Psi, S_list, ns))
print(ans3 <- correlateR:::rcm_get_nu_nlm(Psi, S_list, ns))
print(ans4 <- correlateR:::rcm_get_nu_optimize2(Psi, S_list, ns))

abline(v = ans$maximum, col = "orange", lwd = 2, lty = 2)
abline(v = ans2$par, col = "blue", lwd = 2, lty = 3)
abline(v = ans3$estimate, col = "brown", lwd = 2, lty = 4)
abline(v = ans4$maximum, col = "purple", lwd = 2, lty = 4)

\dontrun{
library("microbenchmark")
microbenchmark(correlateR:::rcm_get_nu_optimize2(Psi, S_list, ns),
               correlateR:::rcm_get_nu_optimize(Psi, S_list, ns),
               correlateR:::rcm_get_nu_optim(Psi, S_list, ns),
               correlateR:::rcm_get_nu_nlm(Psi, S_list, ns))
}
}
\author{
Anders Ellern Bilgrau
}
\keyword{internal}

