% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcm_loglik_arma}
\alias{rcm_loglik_arma}
\title{The RCM log-likelihood function}
\usage{
rcm_loglik_arma(Psi, nu, S_list, ns)
}
\arguments{
\item{Psi}{A \code{p} times \code{p} numeric positive semi-definte matrix.}

\item{nu}{A numeric of length 1 giving the degrees of freedom.}

\item{S_list}{A \code{list} of scatter matrices of the same size as
\code{Psi} for each group.}

\item{ns}{A numeric of the same length as \code{S_list} giving the
number of samples in each group.}
}
\value{
The value of the log-likelihood.
}
\description{
Fast evaluation of the RCM log-likelihood function.
}
\examples{
ns <-  c(5, 5, 5)
S <- createRCMData(ns = ns, psi = diag(4), nu = 30)
correlateR:::rcm_loglik_arma(Psi = diag(4), nu = 15, S_list = S, ns = ns)
}
\author{
Anders Ellern Bilgrau <anders.ellern.bilgrau (at) gmail.com>
}
\keyword{internal}

