% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rwishart}
\alias{rinvwishart}
\alias{rwishart}
\title{Wishart and inverse-Wishart distributions}
\usage{
rwishart(n, sigma = diag(3), nu = 10)

rinvwishart(n, psi = diag(3), nu = 10)
}
\arguments{
\item{n}{The number of realizations.}

\item{sigma}{A \code{p} by \code{p} symmetric postive-definite matrix.}

\item{nu}{An numeric of length one giving the degrees of freedom.}

\item{psi}{A square symmetric postive-definite matrix.}
}
\value{
\code{rwishart} returns an \code{p} by \code{p} by \code{n}
  3-dimensional array where each slice is a draw from the Wishart
  distribution.

For \code{rinvwishart} each slice is a draw from the inverse Wishart
  distribution.
}
\description{
Generate data from Wishart and inverse-Wishart distributions using the
Barlett decomposition.
}
\examples{
rwishart(n = 2)
rwishart(n = 3, diag(4), nu = 4)
rinvwishart(n = 2)
}
\seealso{
\code{\link{rWishart}} from the \pkg{stats}-package.
}

