% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wishart.R
\name{rwishart}
\alias{rwishart}
\alias{rinvwishart}
\title{Wishart and inverse-Wishart distributions}
\usage{
rwishart(n, sigma = diag(3), nu = 10, verbose = TRUE)

rinvwishart(n, psi = diag(3), nu = 10)
}
\arguments{
\item{n}{The number of realizations.}

\item{sigma}{A \code{p} by \code{p} symmetric and positive definite matrix.}

\item{nu}{An \code{numeric} of length one giving the degrees of freedom.}

\item{verbose}{logical. Warn if draws are from a singular Wishart 
distribution.}

\item{psi}{A square symmetric positive definite matrix.}
}
\value{
\code{rwishart} returns an \code{p} by \code{p} by \code{n} 
  3-dimensional array where each slice is a draw from the Wishart 
  distribution.

For \code{rinvwishart} each slice is a draw from the inverse Wishart 
  distribution.
}
\description{
Generate data from Wishart and inverse-Wishart distributions using the 
Barlett decomposition.
When simulating from the Wishart distribution and \code{nu} is an integer
strictly less than p, then the functions draws from a singular Wishart 
distribution using multivariate normal draws.
}
\examples{
rwishart(n = 2)
rwishart(n = 3, diag(4), nu = 4)

rwishart(n = 1, diag(3), nu = 3)
rwishart(n = 1, diag(3), nu = 2)
rwishart(n = 1, diag(3), nu = 1)
rinvwishart(n = 2)
}
\seealso{
\code{\link{rWishart}} from the \pkg{stats}-package.
}
