% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rwishart}
\alias{rinvwishart}
\alias{rwishart}
\title{Wishart and inverse-Wishart distributions}
\usage{
rwishart(n, sigma = diag(3), nu = 10L)

rinvwishart(n, Psi = diag(3), nu = 10L)
}
\arguments{
\item{n}{The number of realizations.}

\item{sigma}{A square symmetric postive-definite matrix.}

\item{nu}{An integer giving the degrees of freedom.}

\item{Psi}{A square symmetric postive-definite matrix.}
}
\value{
\code{rwishart} returns a matrix drawn from the Wishart distribution.

\code{rinvwishart} returns a matrix drawn from the inverse
  Wishart distribution.
}
\description{
Generate data from a Wishart and inverse-Wishart distribution.
}
\examples{
rwishart(n = 2)
rWishart(n = 2, df = 10L, Sigma = diag(3))
rwishart(n = 3, diag(4), nu = 4)
n <- 10000
a <- replicate(n,diag(3)) - rwishart(n)/10L
hist(a)
rinvwishart(n = 2)
}
\seealso{
\code{\link{rWishart}} from the \pkg{stats}-package.
}

