% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scatter}
\alias{scatter}
\alias{xscatter}
\title{Compute the scatter matrix}
\usage{
scatter(X)

xscatter(X, Y)
}
\arguments{
\item{X}{An \code{n} by \code{p1} numeric matrix with observations in the rows
and variables in the columns.}

\item{Y}{An \code{n} by \code{p2} as \code{X}.}
}
\value{
The scatter matrix or cross-scatter matrix.
}
\description{
This functions is a fast computation of the scatter matrix and cross-scatter
matrix given by
  \deqn{\sum_{i = 1}^n(x_i-\mu)(x_i-\mu)^T}{XX^T} and \eqn{XY^T}{XY^T}
respectively.
}
\examples{
n <- 11
X <- createData(n, 4)
Y <- createData(n, 2)
scatter(X)
(n-1)*cov(X)

xscatter(X, Y)
n*xcov(X, Y, method = "ML")  # == (n - 1)*xcov(X, Y)
}
\author{
Anders Ellern Bilgrau <abilgrau (at) math.aau.dk>
}
\keyword{internal}

