% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Scor.R, R/Scov.R
\name{Scor}
\alias{Scor}
\alias{Scov}
\title{Shrinkage covariance estimation}
\usage{
Scor(X, method = c("OAS", "RBLW", "LW", "SS"))

Scov(X, method = c("OAS", "RBLW", "LW", "SS"))
}
\arguments{
\item{X}{The data matrix of size \code{n} by \code{p}.}

\item{method}{The method of estimating the optimal interpolating parameter.
The default is OAS.}
}
\value{
A \code{p} by \code{p} numeric matrix with two extra attributes giving
  the used mixture (\eqn{\rho}{rho}) and the method.
}
\description{
This algorithm seeks to find a covariance (dense) estimate
that (asymptotically) minimizes the mean-squared error (MSE) obtained by
linear shrinkage problem as proposed by Ledoit and Wolf (LW).
It is effectively a interpolation/mix of the sample ML
estimate of the covariance matrix, \eqn{S}, and the most well-conditioned
(and naive) estimate \eqn{F = 1/p tr(S) I}.
}
\details{
The improved estimate using Rao-Blackwell theorem, abbreviated RBLW, and
  the oracle approximating shrinkage (OAS) are also implemented. The
  algorithm seeks a solution to the problem:
    \deqn{\min_{\rho}E[ || \Sigma_O - \Sigma ||^2 ]}{%
          minimize E[ || Sigma_O - Sigma ||^2 ] w.r.t. rho}
    \deqn{s.t. \Sigma_O = (1 - \rho)S + \rho F}{%
          s.t. Sigma_O = (1-rho)*S + rho*F}
  using various methods
  The interpolated \eqn{\rho}{rho} value used is always
  \eqn{\min(\rho, 1)}{min(rho,1)}.
  More information can be found in the given reference.
}
\examples{
n <- 3
X <- createData(n, 5)
cov(X)
Scov(X, method = "OAS")
Scov(X, method = "RBLW")
Scov(X, method = "LW")
}
\references{
Ledoit, O., & Wolf, M. (2004). A well-conditioned estimator for
  large-dimensional covariance matrices. Journal of Multivariate Analysis,
  88(2), 365-411. doi:10.1016/S0047-259X(03)00096-4

  Chen, Y., & Wiesel, A. (2010). Shrinkage algorithms for MMSE covariance
  estimation. Signal Processing, IEEE, 58(734), 1-28.
  Methodology; Computation.
  \url{http://arxiv.org/abs/0907.4698}

  Schafer, J., & Strimmer, K. (2005). A shrinkage approach to large-scale
  covariance matrix estimation and implications for functional genomics.
  Statistical Applications in Genetics and Molecular Biology, 4(1).
  \url{http://www.stat.wisc.edu/courses/st992-newton/smmb/files/expression/shrinkcov2005.pdf}
}

