% Generated by roxygen2 (4.0.2): do not edit by hand
\name{soft}
\alias{soft}
\title{The soft thresholding function}
\usage{
soft(a, c)
}
\arguments{
\item{a}{A vector/matrix to apply the soft thresholding to.}

\item{c}{A numeric giving the soft threshold.}
}
\value{
A vector/matrix the same size as \code{a}.
}
\description{
The soft thresholding function given by
\deqn{soft(a, c) = sgn(a)\cdot(|a| - c)_+}{
      soft(a, c) = sgn(a)*max(|a| - c, 0)}
}
\examples{
a <- rnorm(100)
soft(a, c = 0.2)

x <- createData(n = 10, m = 4)
soft(x, c = 0.7)
}

